backend/requirements.txt
fastapi
fastapi[standard]
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
pydantic
python-multipart
mysqlclient
httpx
bcrypt==4.0.1
Pillow>=9.0.0